GP_SERV_COMMAND_REGISTRATION
