+++ src/map/packets/s2c/0x0bf_registration.h
@@ -1,0 +1,43 @@
+/*
+===========================================================================
+
+  Copyright (c) 2025 LandSandBoat Dev Teams
+
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see http://www.gnu.org/licenses/
+
+===========================================================================
+*/
+
+#pragma once
+
+#include "base.h"
+
+class CBaseEntity;
+
+// https://github.com/atom0s/XiPackets/tree/main/world/server/0x00BF
+// This packet is sent by the server to update a client's event information when interacting with a battlefield registration NPC.
+class GP_SERV_COMMAND_REGISTRATION final : public GP_SERV_PACKET<PacketS2C::GP_SERV_COMMAND_REGISTRATION, GP_SERV_COMMAND_REGISTRATION>
+{
+public:
+    struct PacketData
+    {
+        uint16_t unknown04;     // PS2: (New; did not exist.)
+        uint16_t Result;        // PS2: (New; did not exist.)
+        uint32_t unknown08;     // PS2: (New; did not exist.)
+        uint32_t ActIndex;      // PS2: (New; did not exist.)
+        uint8_t  padding10[12]; // PS2: (New; did not exist.)
+    };
+
+    GP_SERV_COMMAND_REGISTRATION(const CBaseEntity* PEntrance, uint32_t response);
+};
